Optimize Library generation
Templates now get a header JSON string in the first line.
It is marked with #! and is left out from the library file
For now only the number of Pins is in the header and
needs to fit the parts solder joint field.
Preperations are put into the generator function to
have part specific templates and the group template as
fall back if a part tempate is not found.

Template files are updated to contain pins header information.